requirements.txt
pandas
numpy
matplotlib
seaborn


python-dotenv
certifi
scikit-learn
dvc
dagshub
dvc-s3
mlflow
scipy

python-whois
bs4
streamlit

geopy   
apache-airflow